#25 feat:메인페이지 레이아웃 생성
