+++ src/features/main/mainpage/MainPageEarth.css
@@ -23,27 +23,8 @@
   border-radius: 50%;
   transform-style: preserve-3d;
 
-  /* background: white; */
-
   animation: spin 30s linear infinite;
 }
-
-/* 위도선 */
-/* .latitude {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 400px;
-  height: 400px;
-
-  border-radius: 50%;
-  border: 2px solid rgba(214, 14, 14, 0.5);
-  background: linear-gradient(85deg, #f8830e -33.87%, #427e9c 81.79%);
-
-  transform-style: preserve-3d;
-  transform-origin: center;
-  translate: -50% -50%;
-} */
 
 .latitude2 {
   position: absolute;
