+++ src/features/main/mainpage/MainPageEarth.tsx
@@ -1,60 +1,13 @@
 import React from "react";
-import "./TestEarth.css";
+import "./MainPageEarth.css";
 
-const LAT_LINES = 3; // 위도선 개수
 const LNG_LINES = 10; // 경도선 개수
 
 const EarthSphere: React.FC = () => {
   return (
     <div className="scene">
-      {/* 자전축 */}
       <div className="axis">
         <div className="sphere">
-          {/* 위도 */}
-          {/* {Array.from({ length: LAT_LINES }).map((_, i) => {
-            const angle = (360 / LAT_LINES) * i;
-
-            return (
-              <div
-                key={`lat-${i}`}
-                className="latitude"
-                style={{ transform: `rotateX(${angle}deg)` }}
-              />
-            );
-          })} */}
-
-          {/* {Array.from({ length: LAT_LINES }).map((_, i) => {
-            const angle = (360 / LAT_LINES) * i;
-
-            return (
-              <div
-                key={`lat-${i}`}
-                className="latitude2"
-                style={{
-                  // transform: `rotateX(${angle}deg)`,
-                  width: `${(i + 1) * 180}px`,
-                  height: `${(i + 1) * 180}px`,
-                }}
-              />
-            );
-          })} */}
-
-          {/* {Array.from({ length: LAT_LINES }).map((_, i) => {
-            const angle = (360 / LAT_LINES) * i;
-
-            return (
-              <div
-                key={`lat-${i}`}
-                className="latitude2"
-                style={{
-                  // transform: `rotateX(${angle}deg)`,
-                  width: `${(i + 1) * 180}px`,
-                  height: `${(i + 1) * 180}px`,
-                }}
-              />
-            );
-          })} */}
-
           <div
             className="latitude2"
             style={{
