+++ src/features/main/mainpage/MainPageLattice.css
@@ -1,0 +1,58 @@
+.LatticeFrame {
+  position: absolute;
+  top: 0%;
+  left: 0%;
+  width: 100%;
+  height: 100%;
+  display: grid;
+  place-items: center;
+}
+
+.BoxOne {
+  width: 100%;
+  height: 100%;
+  grid-area: 1 / 1;
+
+  background: linear-gradient(
+    90deg,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 0) 35%,
+    #ffffff 50%,
+    rgba(255, 255, 255, 0) 85%,
+    rgba(255, 255, 255, 0) 100%
+  );
+}
+
+.BoxTwo {
+  width: 100%;
+  height: 100%;
+  grid-area: 1 / 1;
+
+  background: linear-gradient(
+    180deg,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 0) 35%,
+    #ffffff 50%,
+    rgba(255, 255, 255, 0) 85%,
+    rgba(255, 255, 255, 0) 100%
+  );
+}
+
+.box {
+  position: absolute;
+  top: 0%;
+  left: 0%;
+  width: 100%; /* 네모난 상자 */
+  height: 100%;
+  background-color: black;
+
+  /* 타원형 퍼지는 그라데이션 */
+  background-image: radial-gradient(
+    ellipse 70% 40% at center,
+    rgba(255, 255, 255, 0.9) 0%,
+    rgba(255, 255, 255, 0.6) 25%,
+    rgba(255, 255, 255, 0.3) 45%,
+    rgba(255, 255, 255, 0.1) 60%,
+    rgba(255, 255, 255, 0) 75%
+  );
+}
