+++ src/features/main/mainpage/MainPageLattice.tsx
@@ -1,0 +1,13 @@
+import "./MainPageLattice.css";
+
+const Lattic = () => {
+  return (
+    <div className="box"></div>
+    // <div className="LatticeFrame">
+    //   <div className="BoxOne"></div>
+    //   <div className="BoxTwo"></div>
+    // </div>
+  );
+};
+
+export default Lattic;
