+++ src/features/main/mainpage/MainPageStars.css
@@ -10,16 +10,13 @@
   }
 }
 
-/* 전체 배경 */
 .star-container {
   width: 100%;
   height: 100vh;
-  background: black;
   overflow: hidden;
   position: absolute;
 }
 
-/* 별 기본 스타일 */
 .star {
   position: absolute;
   background-color: #f5f5f5;
