--- src/features/main/pages/TestFuck.css
@@ -1,28 +1,0 @@
-.container {
-  width: 150px;
-  height: 150px;
-  border: 4px solid blanchedalmond;
-  margin: 100px;
-  perspective: 300px; /* 자식 요소에 원근 효과 적용 */
-  /* transform: perspective(100px) rotateX(0deg); */
-  /* perspective 함수는 자신에게 정의 */
-  transform-style: preserve-3d; /* 자식 요소에도 3D 효과를 보존함 (단 후손은 적용 안 됨) */
-}
-.container .item {
-  width: 150px;
-  height: 150px;
-  background-color: orange;
-  /* transform: rotateY(40deg); */
-  /* transform: rotateX(13deg); */
-  transform: rotateX(-50deg);
-  transform: rotateZ(45deg);
-
-  transform-style: preserve-3d; /* 부모 요소에서 작성했지만 자식에만 해당하여 한 번 더 작성해야 함 */
-}
-
-.container .item .box {
-  width: 100px;
-  height: 100px;
-  background-color: royalblue;
-  transform: rotateX(43deg);
-}
