--- src/features/main/pages/TestFuck.tsx
@@ -1,11 +1,0 @@
-import "./TestFuck.css";
-
-const Funk = () => {
-  return (
-    <div className="container">
-      <div className="item">{/* <div className="box"></div> */}</div>
-    </div>
-  );
-};
-
-export default Funk;
