--- src/features/main/pages/main.css
@@ -1,159 +1,0 @@
-.MainPageFrame {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: black;
-  height: 4500px;
-  width: 1920px;
-}
-
-.MainPageFristSection {
-  width: 1920px;
-  height: 1237px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.DistrictImgFrame {
-  margin-top: 120px;
-  margin-bottom: 200px;
-}
-
-.MainPageSpaceOne {
-  width: 1920px;
-  height: 200px;
-}
-
-.MovingRightBox {
-  width: 461px;
-  height: 312px;
-  border-radius: 44px;
-
-  background:
-    linear-gradient(#000, #000) padding-box,
-    linear-gradient(0deg, #ffffff 9.86%, #76cbf6 74.84%) border-box;
-
-  border: 1px solid transparent;
-
-  box-shadow: 0px 5px 20px 4px #9ed8f587;
-
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  animation: moveBoxRight 16s linear infinite;
-}
-
-@keyframes moveBoxRight {
-  0% {
-    transform: translateX(-461px);
-  }
-  100% {
-    transform: translateX(calc(1920px + 461px));
-  }
-}
-
-.MovingLeftItem {
-  width: 461px;
-  height: 312px;
-  border-radius: 44px;
-
-  background:
-    linear-gradient(#000, #000) padding-box,
-    linear-gradient(0deg, #ffffff 9.86%, #76cbf6 74.84%) border-box;
-
-  border: 1px solid transparent;
-
-  box-shadow: 0px 5px 20px 4px #9ed8f587;
-
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  animation: moveBoxLeft 16s linear infinite;
-}
-
-@keyframes moveBoxLeft {
-  0% {
-    transform: translateX(calc(1920px + 461px));
-  }
-  100% {
-    transform: translateX(-461px);
-  }
-}
-
-.MovingItemContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 262px;
-  width: 450px;
-  position: relative;
-}
-
-.MovingItemTitle {
-  position: absolute;
-  top: 0px;
-
-  font-family: Pretendard;
-  font-weight: 600;
-  font-style: SemiBold;
-  font-size: 36px;
-  leading-trim: NONE;
-  line-height: 100%;
-  letter-spacing: -3%;
-
-  background: linear-gradient(
-    178.98deg,
-    #62e1ef -4.89%,
-    #7ac7f7 18.12%,
-    #92adff 46.03%
-  );
-
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.MovingItemSub {
-  position: absolute;
-  bottom: 0px;
-
-  font-family: Inter;
-  font-weight: 600;
-  font-style: Semi Bold;
-  font-size: 14px;
-  leading-trim: NONE;
-  line-height: 18px;
-  letter-spacing: -3%;
-}
-
-.PartIntroduceSection {
-  width: 1920px;
-}
-
-.PartIntroduceSectionContainer {
-  width: 1920px;
-  height: 446px;
-  margin-top: 120px;
-
-  display: flex;
-  justify-content: space-around;
-}
-
-.PartIntroduceComponentFrame {
-  width: 333px;
-  height: 446px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.MainPageSpaceTwo {
-  width: 1920px;
-  height: 200px;
-}
