+++ src/routes/AppRouter.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import PendingPage from "../shared/pages/PendingPage";
 import PartMainPage from "../features/partpage/pages/PartMainPage";
-import MainPage from "../features/main/pages/MainPage";
+import MainPage from "../features/main/mainpage/MainPage";
 import RecruitPage from "../features/recruit/pages/RecruitPage";
 import ProjectDetailPage from "../features/project/pages/ProjectPage";
 import ApplyNoticePage from "../features/apply/ApplyPage";
